refactor fastapi server for split into dependencies and monzo_api(both auth and monzo apis) and singleton pattern for redis connection, reconstruct refresh_token handler to refresh_tokens_endpoint() and perform_token_refresh()
